thesis_check/requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
Django==5.2.7
djangorestframework==3.15.1
en-core-web-sm==3.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.4
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==1.26.4
packaging==25.0
preshed==3.0.10
psutil==7.1.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.32.5
rich==14.2.0
scikit-learn==1.5.2
scipy==1.15.3
shellingham==1.5.4
smart_open==7.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
